refactor: move showReportProblem
