+++ app/routes/shared/vorabcheck.server.ts
@@ -30,7 +30,7 @@ import { validateFormData } from "~/services/validation/validateFormData.server"
 import { applyStringReplacement } from "~/util/applyStringReplacement";
 import { getButtonNavigationProps } from "~/util/buttonProps";
 import { filterFormData } from "~/util/filterFormData";
-import { shouldShowReportProblem } from "./showReportProblem";
+import { shouldShowReportProblem } from "../../components/reportProblem/showReportProblem";
 
 export const loader = async ({
   params,
